LayoutTests/accessibility/mac/slider-supports-actions-expected.txt

X
This test makes sure that AXIncrement and Decrement are exposed on sliders (ARIA and input).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS succeeded is true
PASS succeeded is true
PASS succeeded is true
PASS succeeded is true
PASS succeeded is true
PASS succeeded is true
PASS successfullyParsed is true

TEST COMPLETE
